requirements.txt
configparser==4.0.2
pylint
mysql-connector>=2.2.9
psutil>=5.6.7
requests>=2.22.0